refactor(routes): split 1900-line collection.js into focused route modules
collection.js had grown to ~1950 lines spanning unrelated concerns. Split by
responsibility so each area is a small, ownable file; pure relocation, verified
identical by the 32-case e2e suite.

- utils/collectionHelpers.js: shared helpers (compartment resolve/describe,
  checked-out allocation, default plan, rule-config normalize) — one neutral
  home so route files never import each other
- routes/storage.js: locations + compartments + recommend/apply/resort
- routes/stats.js: /stats, /stats/history, /cards/:id/price-history
- routes/importExport.js: /export, /import
- collection.js: now just search/scan + collection CRUD (543 lines)
- server.js: mount the new routers at /api; trim now-unused imports
