+++ backend/src/server.js
@@ -12,6 +12,9 @@ const authRoutes = require('./routes/auth');
 const sharedRoutes = require('./routes/shared');
 const adminRoutes = require('./routes/admin');
 const collectionRoutes = require('./routes/collection');
+const storageRoutes = require('./routes/storage');
+const statsRoutes = require('./routes/stats');
+const importExportRoutes = require('./routes/importExport');
 const setsRoutes = require('./routes/sets');
 const decksRoutes = require('./routes/decks');
 const settingsRoutes = require('./routes/settings');
@@ -168,6 +171,9 @@ app.use('/api/auth', authRoutes);
 app.use('/api/shared', sharedRoutes);
 app.use('/api/admin', adminRoutes);
 app.use('/api', collectionRoutes);
+app.use('/api', storageRoutes);
+app.use('/api', statsRoutes);
+app.use('/api', importExportRoutes);
 app.use('/api/sets', setsRoutes);
 app.use('/api/decks', decksRoutes);
 app.use('/api/settings', settingsRoutes);
